Work on the "header" block (2).
html - Work on the "header" block has started. Basic tags and classes for working with the block were created. The code structure of the block was worked out.

css/scss - Basic site parameters were applied for html and body tags. The "header" block was added to the file. A section with common site classes was added.