Progress on the "footer" block (3).
html - Content and additional tagging was added to the main tags and classes. In the process, new generic classes were created for the content. The block structure was redesigned.

css/scss - During the work were added and worked through the basic properties of css styles, positioning, flex-elements and various animations, for the block "footer" and its classes and tags. New general classes were added to the scss file.

img - an image was added, to work with the block.