Progress on the "productivity" block (4).
html - additional tags were added and a small part of the code was slightly reworked.

css/scss - The import of fonts was changed, the font "Roboto" was added. Variables were created to work with fonts. In the course of work block "productivity__body" and its classes, got properties of css styles, flex-elements, positioning and animation, when working with the interactive elements of the site.

img - images were added to work with the block.