Progress on the "about" block (7).
html - a new block "about__stats" was created. Two main classes were added to it and subclasses were added.

css/scss - the "about__stats" block was added to the scss/css file. The main parameters of the block were added.

img - images have been added to work with the block.