Progress on the "testimonials" block (2).
html - Content was added to the "testimonials" block, its classes, subclasses and tags also received content. The code structure was reworked and additional classes were added.

css/scss - The testimonials block, has been added to the scss/css file. Its main block got basic style properties and flex elements. To the general block, a script from js was added. The structure of the common classes was redesigned.

img - an image was added to work with the block.

js - the "slider.js" script was created. It is designed for round slider buttons, when you click on them, the buttons are colored and remain so, until you click on others.