Progress on the "help" block (4).
html - The code structure was reworked once again.

css/scss - Smoothness was added for future animations, animations were added for the button when hovering. A general class ".activeList" was added, it appears when you click on the list, with "activeList.js".

js - This script, is responsible for the list in this section, when you click, it opens and the information on the list selected by the user appears. Additional content appears.