Work on the "started-now" block (3) is finished.
html - The library "wow.js" was used, its properties for the main classes and tags were applied. Parameters for animations were added.

css/scss - The scss/css style properties for the main classes and tags of the "started-now" block were applied. Flex and grid elements, positioning, and using pseudoclasses, animations were created when hovering over the content.

img - An image was added to work with the block.